refactor characters to avoid repeating code
